Moved methods taskToString() and fromString() from FileBackedTaskManager to TaskToStringConverter
